Fix style matching in Firefox 67
(Ported from PR google/blockly#2485, which is by @NeilFraser.
 Description from that PR follows:)

Chrome returns ‘transform: translate(107px, 0px);’ whereas Firefox now
returns ‘transform: translate(107px);’ if the y value is 0.  This is
consistent with existing behaviour in the translate SVG attribute.

The comment in our code specifically states:
// Accounts for same exceptions as XY_REGEX_
Yet that was not true at all.

This PR makes the y argument optional (as falsely described in the
comment).  It also merges the 2D and 3D regeps together to simplify the
code.